Convert playback.c to use unsigned byte counters, and _minor_ refactoring of buffer callbacks
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@9227 a1c6a512-1295-4272-9138-f99709370657

+++ apps/playback.h
@@ -35,18 +35,18 @@
 
 #define MAX_TRACK 32
 struct track_info {
-    struct mp3entry id3;     /* TAG metadata */
-    char *codecbuf;          /* Pointer to codec buffer */
-    long codecsize;        /* Codec length in bytes */
+    struct mp3entry id3;       /* TAG metadata */
+    char *codecbuf;            /* Pointer to codec buffer */
+    size_t codecsize;          /* Codec length in bytes */
     
-    off_t filerem;           /* Remaining bytes of file NOT in buffer */
-    off_t filesize;          /* File total length */
-    off_t filepos;           /* Read position of file for next buffer fill */
-    off_t start_pos;         /* Position to first bytes of file in buffer */
-    volatile int available;  /* Available bytes to read from buffer */
-    bool taginfo_ready;      /* Is metadata read */
-    int playlist_offset;     /* File location in playlist */
-    bool event_sent;         /* Has event callback functions been called? */
+    size_t filerem;            /* Remaining bytes of file NOT in buffer */
+    size_t filesize;           /* File total length */
+    size_t filepos;            /* Read position of file for next buffer fill */
+    size_t start_pos;          /* Position to first bytes of file in buffer */
+    volatile size_t available; /* Available bytes to read from buffer */
+    bool taginfo_ready;        /* Is metadata read */
+    int playlist_offset;       /* File location in playlist */
+    bool event_sent;           /* Has event callback functions been called? */
 };
 
 /* Functions */
